app-classic/find_artek/requirements.txt
chardet==2.1.1
Django==1.6.11
# django-auth-ldap==1.1.4
django-extensions==1.2.5
django-olwidget==0.61.0
ftputil==2.7.1
# MySQL-python==1.2.3
numpy==1.9.0
Pillow==1.1
pybtex==0.16
# pysqlite==2.6.3
python-dateutil==2.1
python-ldap==2.4.0
PyYAML==3.10
simplejson==3.0.7
six==1.16.0
sorl-thumbnail==11.12.1b0
South==0.7.6
Unidecode==0.4.10
virtualenv==12.1.1
xlrd==0.9.0
